LayoutTests/imported/w3c/web-platform-tests/largest-contentful-paint/observe-svg-data-uri-image-expected.txt



FAIL Same-origin image is observable. assert_between_inclusive: loadTime should occur between the lower bound and the renderTime expected a number greater than or equal to 60 and less than or equal to 79 but got 0
